test(dart_frog_cli): run e2e tests on os matrix

+++ packages/dart_frog_cli/e2e/test/helpers/kill_dart_frog_server.dart
@@ -1,8 +1,17 @@
 import 'dart:io';
 
-Future<void> killDartFrogServer() async {
-  final result = await Process.run('pkill', ['-f', 'dart_frog']);
-  if (result.exitCode != 0) {
-    throw Exception('pkill -f dart_frog exited with code $exitCode');
+Future<void> killDartFrogServer(int pid) async {
+  if (Platform.isWindows) {
+    final result = await Process.run('taskkill', ['/F', '/T', '/PID', '$pid']);
+    if (result.exitCode != 0) {
+      throw Exception('taskkill /F /T /PID $pid exited with code $exitCode');
+    }
+  }
+
+  if (Platform.isLinux || Platform.isMacOS) {
+    final result = await Process.run('pkill', ['-f', 'dart_frog']);
+    if (result.exitCode != 0) {
+      throw Exception('pkill -f dart_frog exited with code $exitCode');
+    }
   }
 }
